requirements.txt
# Core ML Libraries  
torch>=2.0.0
transformers>=4.36.0
peft>=0.13.2,<0.18.0
accelerate>=0.24.0
datasets>=2.14.0
huggingface-hub>=0.19.0
safetensors>=0.4.0

# Gradio for Hugging Face Spaces
gradio>=4.0.0

# Backend API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Audio Processing (optional)
soundfile>=0.12.1
librosa>=0.10.1

# Utilities
tqdm>=4.65.0
requests>=2.31.0
aiofiles>=23.0.0